LayoutTests/fast/workers/worker-terminate-forever-expected.txt

CONSOLE MESSAGE: line 15: Worker was started
CONSOLE MESSAGE: line 34: Worker was terminated
Test Worker.terminate() for a worker that tries to run forever.
